Update ubuntu.sh
replaces references to 'sudo' with 'sudo -E' so that environment variables, specifically http_proxy and https_proxy, persist to the new shell when cpanm and apt are called. Now, if you are attempting to install CIFv2 from behind a proxy via easybutton, all you should have to do is to make sure your proxy environment variables are set and the script should be able to make successful http requests.